requirements.txt
# Flask Core Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1

# Database Drivers
PyMySQL==1.1.0
SQLAlchemy==2.0.23

# Security & Authentication
Werkzeug==3.0.1
cryptography==41.0.7
bcrypt==4.1.2

# Date/Time Utilities
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0

# PDF Generation (for certificates)
reportlab==4.0.7
PyPDF2==3.0.1

# Excel/CSV Support
openpyxl==3.1.2
pandas==2.1.4

# Email Support
email-validator==2.1.0

# Development Tools (optional)
Flask-DebugToolbar==0.14.1

# Production Server
gunicorn==21.2.0

# Testing (optional)
pytest==7.4.3
pytest-flask==1.3.0

# Code Quality (optional)
flake8==6.1.0
black==23.11.0

# Backup Utilities
schedule==1.2.0

# Greek Language Support
unidecode==1.3.7

# Image Processing (optional)
Pillow==10.1.0

# JSON Handling
simplejson==3.19.2

# HTTP Requests (για external APIs)
requests==2.31.0

# Caching (optional)
Flask-Caching==2.1.0
redis==5.0.1

# Monitoring (optional)
psutil==5.9.6

# File Handling
pathlib2==2.3.7

# Math & Statistics
numpy==1.26.2

# Logging
colorlog==6.8.0

# Additional Production Dependencies
Flask-Limiter==3.5.0
sentry-sdk[flask]==1.38.0
flask-talisman==1.1.0
celery==5.3.4
python-magic==0.4.27
weasyprint==60.2
matplotlib==3.8.2
structlog==23.2.0
alembic==1.13.1
gevent==23.9.1
WTForms==3.1.0
click==8.1.7
itsdangerous==2.1.2
coverage==7.3.2
